Remove `if has("autocmd")`
Neovim will always be compiled with the `autocmd` support and Neovim is
the only version of Vim I ever use. I will not be using vim-tiny or any
other version that has no `autocmd` support.

For this reason, I don't want to sprinkle my "init.vim" with `if
has("autocmd")` to check for `autocmd` support.

See:

 - https://vi.stackexchange.com/a/8568/14805
 - https://vi.stackexchange.com/questions/8566/what-is-the-relevance-of-hasautocmd-today#comment15012_8568